my-backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
psycopg2-binary
Flask_Migrate
PyJWT